Revert "added signup/login component and structure changes"
