+++ src/app/app.component.html
@@ -1,2 +1,50 @@
+<div id="map"></div>
+<div id="transportation-mode">
+  <select id="mode-select" [(ngModel)]="currentMode">
+    <option value="walking">Walking</option>
+    <option value="cycling">Cycling</option>
+    <option value="driving">Driving</option>
+  </select>
+</div>
+
+
+<app-directions
+  [map]="map">
+</app-directions>
+<div id="legend">
+  <h3>Map Legend</h3>
+  <div class="dot-container">
+    <span class="reddot"></span>
+    <p class="text">500-1000 students</p>
+  </div>
+  <div class="dot-container">
+    <span class="greendot"></span>
+    <p class="text">300-500 students</p>
+  </div>
+  <div class="dot-container">
+    <span class="yellowdot"></span>
+    <p class="text">100-300 students</p>
+  </div>
+  <div class="dot-container iconlgnd">
+    <img class="svgicon" src="../assets/icons/wg.svg" />
+    <p class="text">Wohnung</p>
+  </div>
+  <div class="dot-container iconlgnd">
+    <img class="svgicon" src="../assets/icons/dorm.svg" />
+    <p class="text">Student dorm</p>
+  </div>
+  <div class="dot-container iconlgnd">
+    <img class="svgicon" src="../assets/icons/studio.svg" />
+    <p class="text">Studio apartment</p>
+  </div>
+  <div class="dot-container iconlgnd">
+    <img class="svgicon" src="../assets/icons/stores.svg" />
+    <p class="text">Supermarket</p>
+  </div>
+  <div class="dot-container iconlgnd">
+    <img class="svgicon" src="../assets/icons/sports.svg" />
+    <p class="text">Fitness activities</p>
+  </div>
+</div>
 
 <router-outlet></router-outlet>
