+++ src/app/app.component.spec.ts
@@ -14,11 +14,11 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  // it(`should have as title 'hausfinder'`, () => {
-  //   const fixture = TestBed.createComponent(AppComponent);
-  //   const app = fixture.componentInstance;
-  //   expect(app.title).toEqual('hausfinder');
-  // });
+  it(`should have as title 'hausfinder'`, () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual('hausfinder');
+  });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
