+++ src/app/app.module.ts
@@ -6,15 +6,11 @@ import { AppComponent } from './app.component';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { DirectionsComponent } from './directions/directions/directions.component';
-import { SignupComponent } from './signup/signup/signup.component';
-import { MapComponent } from './map/map/map.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    DirectionsComponent,
-    SignupComponent,
-    MapComponent
+    DirectionsComponent
   ],
   imports: [
     BrowserModule,
