--- src/app/signup/signup/signup.component.ts
@@ -1,19 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-signup',
-  templateUrl: './signup.component.html',
-  styleUrls: ['./signup.component.scss']
-})
-export class SignupComponent {
-
-  ngOnInit() {
-
-  }
-
-  onButtonClick() {
-    document.querySelector('.cont')?.classList.toggle('s--signup');
-  }
-
-
-}
