+++ src/index.html
@@ -20,6 +20,5 @@
 
   <body>
     <app-root></app-root>
-    <!-- <app-signup></app-signup> -->
   </body>
 </html>
